+++ tests/assets/class3.ts
@@ -1,0 +1,7 @@
+export class OtherClass<P, S> {}
+
+export class SomeClass extends OtherClass<{}, {}> {
+  public constructor() {
+      super();
+  }
+}