+++ tests/classes.test.ts
@@ -5,8 +5,33 @@ test('should be able to handle exported classes', () => {
   expect(result).toBe(`declare module "test" {
   export class SomeClass {
     constructor(value: string);
-    value: string;
+    private value: string;
     foo(): string;
+  }
+}`);
+});
+
+test('should be able to handle exported classes', () => {
+  const result = runTestFor('class2.ts');
+  expect(result).toBe(`declare module "test" {
+  export class SomeClass {
+    constructor(bar: boolean, value: string);
+    protected bar: boolean;
+    private value: string;
+    foo(): string;
+    private qxz(): boolean;
+    protected name: string;
+  }
+}`);
+});
+
+test.only('should be able to handle exported generic classes', () => {
+  const result = runTestFor('class3.ts');
+  expect(result).toBe(`declare module "test" {
+  export class OtherClass<P, S> {}
+
+  export class SomeClass extends OtherClass<{}, {}> {
+    constructor();
   }
 }`);
 });